fix(mio): downgrade mio dependancy to 0.6.22
The mio dependancy had been on the newer version 0.7.0 but tokio still
depends on version 0.6.22. The main purpose of a mio version of
UnixStream and UnixListner is to be able to support tokio versions so we
have to downgrade to the same version that tokio uses until they
upgrade.

BREAKING CHANGE: Any uses of this crate with mio version 0.7.0 will no
longer work. The resolution of this issue for now is to also downgrade
go version 0.6.22.