chore: add step to CI workflow for MSRV
The tempfile crate depends on the remove_dir_all crate. A recent change
to remove_dir_all increases its MSRV to 1.40.0 (in the change from
version 0.5.2 to version 0.5.3). This casues the build to fail for our
MSRV of 1.39.0. The temporary workaround is to pin the version of
remove_dir_all to 0.5.2. This affects our own tests but does not affect
downstream users (who don't run the tests) because tempfile (and hence
remove_dir_all) are only used in the tests.