fix: remove default features
The mio-fd, net-fd, and tokio-fd features are now opt-in, rather
than opt-out. The change to the CI workflow ensures that the tests
are still run for all features.

BREAKING CHANGE: remove mio-fd, net-fd, and tokio-fd as default features

All features are now opt-in features. Crates that previously relied on these
features will now have to explicty name the features in the dependencies
entry for for "fd-queue" in their "Cargo.toml" files.

Fixes: #5