ci: split CI workflow and the Release workflow
The CI workflow runs on pull requests and runs the test suite on
the matrix of stable, beta, and nightly and on the MSRV (currently
1.39). The Release workflow runs on push's to the master or alpha
branches. As these are protected branches this means the the Release
workflow runs when a pull request is merged into one of these branches.

The Release workflow does not re-run the full test suite that is run
in the CI workflow, but rather just runs the tests on stable Rust. This
is designed to support the current low-velocity change workflow of a
single active pull request at a time but will break down once we start
having multiple active pull requests at a time. At that point the Release
workflow should be extended to test on the same matrix as the CI
workflow. We could also consider using [bors][https://bors.tech/].