fix(release): change process for committing version number
The steps of the semantic release plugins will change the version number
in the Cargo.toml file. The old means of pushing the changes Cargo.toml
file to GitHub was to create a pull request and then attempt to merge it.
This created difficulites with the required status checks.

Change the method of pushing the altered Cargo.toml file to use the
semantic-release/git plugin. This assumes that the sbosnick-bot that
is used to authenticate semantic release is an administrator and that the
required status checks can be by-passed for administrators.