fix(net): change PhantomData in MsgHdr
Change the PhantomData type to be a pair of a libc::iovec slice and a
u8 slice, both mutable, and both with the lifetime declared as a part
of MsgHdr. These are the two things that are logically a part of the
msghdr in MsgHdr through the msg_iov and msg_control pointers and
their corresponding *len fields.