refactor(net): create an owning type for RawFd's
A Fd from iomsg owns the contained RawFd and will close it if it is
dropped before the RawFd has been extacted through IntoRawFd. This
allows the types in the net and iomsg modules to track the ownership
of a RawFd until it can be handed off to calling code in DequeueFd.