test(net): add test for error condition in iomsg
The new test checks that errors in the underlying sendmsg call in iomsg
bubble up from the call to MsgHdr::send().

This test sufaced an error in the implementation of MsgHdr::encode_fds()
if the MsgHdr was created with an empty control buffer and no fds to
encode. Introduced the CMsgMut type to encapsolate the handling of
cmsghdr's as a fix of this error.