test(net): add tests for some edge conditions in iomsg (#25)
Add tests for trying to take the fds a second time after libc::recvmsg has been called,
taking the fds when the cmsg's returned by a call to libc::recvmsg include more that just
passed fds, and attempting to send more fds than will fit in the supplied control
buffer.

Test that errors from libc::sendmsg and libc::recvmsg bubble up through the
MsgHdr::send() and MsgHdr::recv() calls.

One of these tests surfaced an error in the implementation of 
MsgHdr::encode_fds() if the MsgHdr was created with an empty control buffer 
and no fds to encode. Introduce the CMsgMut type to encapsulate the 
handling of cmsghdr's as a fix of this error.