fix(tokio): upgrade to tokio 1.0.0 (#26)
* fix(tokio): upgrade to tokio 1.0.0

Change this crate's tokio::UnixStream and tokio::UnixListener to wrap
the net::UnixStream and net::UnixListener in the tokio (crate) AsyncFd
and then use the AsyncFd to implement AsyncRead and AsyncWrite. This
cuts out the use of the mio module in the tokio module of this crate.

The upgrade to tokio 1.0.0 requires increasing the MSRV to 1.45.0. This
increase in the MRSV removes the need for the "Patch Dependencies" step 
in the ci workflow.

Fixes: #1

BREAKING CHANGE: increase the MRSV to 1.45.0

This change will require upgrading old compilers to at least 1.45.0.