fix(net): change libc dependencies to allow build (#49)
Building the crate as a dependency from a foreign crate would fail with
an error about a missing Debug implementation on libc::msghdr. The same
build error was not reproducable locally.

The fix is to add the "extra_traits" feature to the libc dependency in
this crate. The nix dev-dependency had the "extra_traits" feature
already enabled. cargo's feature unificaton between the dev-dependencies
and the regular dependencies caused the "extra_traits" feature to be
enabled for libc on any local builds of this crate but not on builds
triggered by a dependency in a foreign crate.

Remove the nix dev-dependency to prevent masking this build error (and
because it is no longer necessary). Change the tests that had used nix
types to implement a shared memory type to instead use the std lib File
type (and the tempfile dependency to generate such a File).

Closes: #44