refactor(biqueue): simplify outgoing queue of fds
Change the outgoing queue of fds from an Option<Vec<RawFd>> that is
recreated after each sendmsg call to just a Vec<RawFd> that is drained
on each call but then reused. This may have some performance benefit but
there are no performance tests to show this so this is really just a
simplication of the code.