feat(biqueue): make the queue size generic
Add a SIZE generic const to BiQueue which defaults to
DEFAULT_FD_QUEUE_SIZE (which is 2). This SIZE is used for the length of
the inbound and outbound queues of fd's. It is also used for the size of
the CMsgBuffer used as a part of the calls to recvmsg and sendmsg
wrapped in through MsgHdr.

Current CMsgBuffer is a SmallVec with the size of the inline part of the
vector based on DEFAULT_FD_QUEUE_SIZE (and not on SIZE). Current
limitations in Rust prevent basing tis on SIZE (though that may change
if #![feature(generic_const_exprs)] is stabalized.