feat(biquque): make the queue size generic (#58)
Add a SIZE generic const to BiQueue which defaults to
DEFAULT_FD_QUEUE_SIZE (which is set to 2). This SIZE is used for the
length of the inbound and outbound queues of fd's. It is also used for
the size of the CMsgBuffer used as a part of the calls to recvmsg and
sendmsg wrapped through MsgHdr.

The value of DEFAULT_FD_QUEUE_SIZE is chosen to match the old value of
BiQueue::FD_QUEUE_SIZE to prevent a breaking change.

Currently CMsgBuffer is a SmallVec with the size of the inline part
of the vector based on DEFAULT_FD_QUEUE_SIZE (and not on SIZE).
Current limitations in Rust prevent basing this on SIZE (though that
may change if #![feature(generic_const_exprs)] is stabalized).

This is part of #55.